Converted the ResourceManager over to loading from disk using concrete Resource classes, rather than factories. This allows subclasses to define how they are loaded AND used.
Updated the ResourceManager to store references to Resources that know how to load from disk and use static functions to mimic the behavior previously handled by Factories.
Split Font into a base Font class an a FontTTF that encapsulates the loading from disk using SDL_ttf.
Converted all old ResourceManagers and factories to the new system.
Did a complete update of the Texture code (it NEEDED it).
	Converted it to the new system, using Resource subclasses.
	Removed the myriad of functions used to handle different dimensions in favor of a smaller set of smarter functions.
	Moved around or deleted numerous functions from Texture and TextureManager.
	Put PixelData in its own place.