Fixed up some minro bugs and setup the world to render with a basic grass texture.
Fixed mouse look in the ConceptState.
Made CheckGLErrors throw an exception if a problem was noticed.
Made a few more conversions from std::string to IdType in ResourceManager.
Fixed a bug where I was passing 1 for the height and depth of textures, rather than the appropriate value (Was breaking ALL non-1d textures).