Initial changes to avoid using singletons for all of the resource managers.
None of the managers are singletons anymore.
Updated Framebuffer to take a texture to remove its TextureManager dependency.
Added references to the ResourceGroupManager in all ResourceFactories.
The MHWorld is now more responsible for creating MHObjects.
Added references to the required managers in MHWorld and the MHSceneManager.
Added the resource manager initialization to the DefaultCore.
Added accessors to the managers in MHCore as a temp hack(THIS NEEDS TO GO AWAY).