Removed the last vestiges of the horrible GUI_WIDTH and GUI_HEIGHT macros.
Viewports now call resize on their sources when they are attached.
EventPump now requires access to the window, allowing us to invert the y coords on the mouse events.
Font now takes the window width and height as parameters to 'print' (not sure I like this!).
The UIManager now relies on its resized width/height to place things.