Add Entities to World.
A single dummy is created in world.initialize. World stores RubyEntities
Ruby-side, and C Entities C-side, in an array and a map<string,Entity*>,
respectively.

Passing a sphere for the Entity's model doesn't seem to render anything
currently. Not sure why.

World does some basic Marking of RubyEntity and Entity (tells SceneManager
to delete the Entity) but this is all very preliminary and subject to change
by someone that knows what they're doing.