Complete revamp of the entity/node system!
Everything is now a scene node and the SceneManager now manages everything as such.
Using the scenemanager to manage nodes is based on templates, now, so user created types may be stored in the scenemanager, too.
Stuff is added to the scene by the nodes now (this is great as entities that aren't attached to anything won't be considered).
Dettaching things from nodes is now fully supported.
Fixed a few minor bugs and added some simple features.
Added Renderables, which are used by the RenderQueue instead of Entities.
Objects now add whatever they want to the RenderQueue, be it themselves or debugging objects.