Fix to the picking issue and some minor refactorings.
We now properly invert the distance before lerping distances when doing frustum scaling.
Fixed up some ortho projection calculations to more exactly match the opengl spec.
We now update ViewFrustums immediately when changes are made, rather than just when we set the transformation matrix (this was a bug).
Fixed a "bug" with the RenderContext::setOrtho2D method, where the far plane was behind the near plane.
Fixed an issue in the MH ortho camera where the far plane was behind the near plane.