Refactoring and a fix for the ray cast bug.
Moved ViewFrustum to Base as it is used in testing. I feel like there is an argument for having it there, as well. Need to think about it some more.
Cleaned up Quat/Matrix creation. Having so many constructors was causing weird bugs when conjoined with implicit vector construction.
Made the single value constructor for Vector3 explicit, as it was causing crazyness.
Removed a bunch of worthless, old functions from my Matrix class.
Matrix multiplication of Vector3s now does a persepctive divide with any W component. This fixes the ray casting bug, but also cements Matrix as a graphics specific sort of class. This will also require thought.