This is a huge and incomplete checkin. I'm getting this in, so I can merge the last month's worth of changes.
This represents a complete rewrite of the rendering system, a complete restructuring of the scene management and content code, a broad update to the UI and the terrain code, and supporting changes throughout Moutainhome. It's way massive, and I don't think I can properly describe, in detail, what is going on with this checkin. It's just too big. I'd be writing forever.