Step one to getting things working again.
I disabled the menu state, which Andrew is working on. Currently we go straight for the loading state and into the game state. This change rips the UI everything out of those two states, fixes numerous bindings issues, and sets the framework for eventual rendering. Currently all we get now is a black screen that doesn't crash :P