Fix for texture handling in shaders.
Added some extra safety checking in the ShaderGLSL::setParameter.
Shader enable/disable now happens only in Material, as it should.
Textures will now be disabled when the associated shader is disabled.
Cleaned up how binding textures to shaders happens and set them to be enabled as they're attached to a shader.