Fixed the depth test bug and enabled wireframe toggling in GameState.
The depth test bug was actually one that loaded materials were always setting their render params, even if they were set explicitely (which is wrong, we want them unset so they default to their parent's value). Additionally, the depth test value was defaulting to 0, when it should be 1, by convention.
The wireframe toggle was as simple as writing a new binding function to handle it.