Revamp of how we handle textures and lighting for the terrain.
We now make use of a texture grid for the terrain, rather than discrete textures. This gives us a performance win and the ability to make use of mincraft tile packs.
Lighting is now a little brighter, and controllable for the terrain via the descriptions file.
Moved some logic in the logging macros to allow for more control over logging.
Normalized some of the descriptor functions so that the the class attr and the instance attr stuff acts the same.
Moved the code that generates a Material for a given Tile into the MHTerrainBindings. This allows us more flexibility in material creation as we no longer need to store all of the information needed to generate the material in the C side tile class; we can just grab it straight out of the ruby class. This also lets us simplify the creation of the c side tile class when setting tiles.
Added in a new shader which is capable of handling tiles with different side/top/bottom textures and added support for accessing this functionality to the description code and in the material generation code (now in the terrain bindings).