Fixed the texture seam issue we were seeing.
We're back to splitting up the the textures into pieces. This might be a little slower, but we're limited elsewhere currently, so it shouldn't be noticeable.
Added support for changing the current texture set, as well, which should leave it fairly simple to change all of our textures at once.