Big rework/cleanup of mountainhome.rb and some terrain bindings.
Normailzed the method names we use for setting parameters with those for class and instance attribs.
Normalized the names we use for class and instance attribs with each other.
Changed parameters to use a hash instead of an array (why were they an array?).
Setup a new system for managing DSL extensions. There is a fixed format for a module that can be used to dynamically extend the DSL. This allows us to put everything associated with a DSL feature in one place, rather than scattering it everywhere. This was the major part of this change, though it wasn't at all the original intent of the change.
Setup the ruby side tile manipulation methods to use tile types rather than tile instances.
Fixed some issues where modules were being stored in 'klass' variables (sorry Paul!).
Fixed a bug where we were mining with the terrain's set_tile method, rather than the world's (fucks with pathfinding).

Note: I pushed this change out quite quickly. I see several things that will need revisiting for style / correctness. This works well enough for now, though.
    Comments in MHTerrainBindings.
    Use of klass should sometimes be new_module.    Not sure I like using 'extends'. Maybe 'language_extension' to avoid confusion with ruby's extend (we're not extending extensions and we're not extending the module when we call module.extends. I just left it like this because it's what it already was.
    There are a bunch of FIXMEs in the new code.
    I would like to remove the terrain accessor from world if we're going to be forcing people to use world's methods for terrain interaction.