Don't rendering terrain groups that don't exist.
Cases that resulted in the removal of a terrain chunk renderable group weren't actually affecting the render op, so visually nothing changed. This caused mined things to not disappear (as mining would remove all tiles in the selected set) and selections to not go away until you selected something else in that same group. To fix this, we simply zero out the index buffer size.