Add fog to the terrain shader.
Refactor handling of cameras, to give control of the far plane to the MHWorld.
Add a constant "view distance" factor to the world that is used to set Camera far planes and for fog calculations.
Enable fog based on the MHWorld's "view distance" value to the terrain and multi-texture terrain shaders.